requirements.txt
pyserial~=3.5
Pillow~=9.4.0
numpy~=1.23.4
tensorflow~=2.12.0
requests~=2.26.0